perf(dashboard): server shell — one render instead of 19 client requests (#1096)
The page is now a server component: the per-user Supabase burst runs on the
cookie-bound server client, content lookups are direct bundle imports instead
of /api/content/* fetches, and review/quests/cohort call their lib functions
directly. Independent sections stream behind Suspense; toasts/celebrations
stay in thin client islands because their dedupe lives in sessionStorage.
