+++ apps/web/src/components/dashboard/review-strip.tsx
@@ -1,14 +1,15 @@
 "use client";
 
-import { useEffect, useState } from "react";
 import Link from "next/link";
 import { useTranslations, useLocale } from "next-intl";
 import { Brain, ArrowRight } from "@phosphor-icons/react";
 import { Button } from "@/components/ui/button";
 
 interface ReviewStripProps {
-  /** Authenticated user id, or null before auth resolves. */
-  userId: string | null;
+  /** Authoritative due count — the session's own length (#977). */
+  count: number;
+  /** Due lesson titles, in session order. */
+  titles: string[];
 }
 
 /** Up to this many lesson titles are named inline before collapsing to "+N". */
@@ -16,52 +17,15 @@ const MAX_TITLES = 3;
 
 /**
  * Dashboard due-review strip (LX-B6). A small additive slot under the Continue
- * card: it names a few due items and deep-links into /review. Reads through
- * `/api/review/due`, which runs the SAME capped `buildReviewSession` as the
- * page (#977: count ≡ what the session serves), and renders nothing when the
- * queue is empty — no empty-state chrome on the dashboard.
+ * card: it names a few due items and deep-links into /review. Presentational
+ * since #1096 — the server shell reads the SAME capped `buildReviewSession` as
+ * the /review page (#977: count ≡ what the session serves) via
+ * `loadReviewDue` and passes the result down; renders nothing when the queue
+ * is empty — no empty-state chrome on the dashboard.
  */
-export function ReviewStrip({ userId }: ReviewStripProps) {
+export function ReviewStrip({ count, titles }: ReviewStripProps) {
   const t = useTranslations("dashboard");
   const locale = useLocale();
-  const [titles, setTitles] = useState<string[]>([]);
-  const [count, setCount] = useState(0);
-
-  useEffect(() => {
-    if (!userId) return;
-    let active = true;
-    (async () => {
-      try {
-        // One authoritative read (#977): /api/review/due runs the SAME
-        // buildReviewSession the /review page renders, so this badge can
-        // never promise an item the session then drops. The previous
-        // client-side intersection with lessons-summary used a weaker rule
-        // (lesson-in-bundle, no owning-course check) and could over-count.
-        const response = await fetch("/api/review/due");
-        if (!active) return;
-        if (!response.ok) {
-          setCount(0);
-          setTitles([]);
-          return;
-        }
-        const data = (await response.json()) as {
-          count?: number;
-          titles?: string[];
-        };
-        if (!active) return;
-        setTitles(Array.isArray(data.titles) ? data.titles : []);
-        setCount(typeof data.count === "number" ? data.count : 0);
-      } catch {
-        if (active) {
-          setCount(0);
-          setTitles([]);
-        }
-      }
-    })();
-    return () => {
-      active = false;
-    };
-  }, [userId]);
 
   // Empty queue → no strip (LX-B6: hidden when the queue is empty).
   if (count === 0) return null;
